Windows portability
git-svn-id: svn+ssh://192.168.1.105/Users/mikael/icl-repository/trunk@671 763894fe-7313-0410-9b4e-9763914fbcec

+++ include/ic_base_header.h
@@ -98,7 +98,7 @@ typedef guint64 IC_CONF_STATE_TYPE;
 struct ic_iovec
 {
   gchar *iov_base;
-  guint64 iov_len;
+  guint32 iov_len;
 };
 
 typedef struct ic_iovec IC_IOVEC;
